refactor: move server setup to main.go and add prometheus metrics
